chore: Convert remaining use of old artist api functions & queries
