+++ mobile/src/modules/scanning/helpers/audio.ts
@@ -13,8 +13,8 @@ import { invalidTracks, tracksToArtists, tracksToGenres } from "~/db/schema";
 
 import { upsertAlbums } from "~/api/album";
 import { AlbumArtistsKey } from "~/api/album.utils";
-import { createArtists } from "~/api/artist";
 import { deleteTracks, upsertTracks } from "~/api/track";
+import { createArtists } from "~/data/artist/api";
 import { createGenres } from "~/data/genre/api";
 import { preferenceStore } from "~/stores/Preference/store";
 import { scanningProgressStore } from "../ScanningProgress";
