+++ mobile/src/navigation/screens/tracks/ModifyView.tsx
@@ -16,9 +16,9 @@ import i18next from "~/modules/i18n";
 import { Info } from "~/resources/icons/Info";
 import { upsertAlbums } from "~/api/album";
 import { AlbumArtistsKey } from "~/api/album.utils";
-import { createArtists } from "~/api/artist";
 import { updateTrack } from "~/api/track";
 import { useTrack, useTrackGenres } from "~/queries/track";
+import { createArtists } from "~/data/artist/api";
 import { createGenres } from "~/data/genre/api";
 import { Resynchronize } from "~/stores/Playback/actions";
 import { usePreferenceStore } from "~/stores/Preference/store";
