--- mobile/src/queries/artist.ts
@@ -1,28 +1,0 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-
-import type { artists } from "~/db/schema";
-
-import { updateArtist } from "~/api/artist";
-import { queries as q } from "./keyStore";
-
-//#region Queries
-/** Get specified artist. */
-export function useArtist(artistName: string) {
-  return useQuery({ ...q.artists.detail(artistName) });
-}
-//#endregion
-
-//#region Mutations
-/** Update specified artist. */
-export function useUpdateArtist(artistName: string) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (
-      updatedValues: Partial<Omit<typeof artists.$inferInsert, "name">>,
-    ) => updateArtist(artistName, updatedValues),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: q.artists._def });
-    },
-  });
-}
-//#endregion
